@@ -1,7 +1,7 @@
 {
   "displayName": "Homebridge dingz Plugin",
   "name": "homebridge-dingz",
-  "version": "1.8.2",
+  "version": "2.0.0-nightly.17",
   "author": "johannrichard",
   "description": "Homebridge Plugin for dingz & myStrom Devices. Implements some (but not all, due to API limitations) functions of a [dingz](https://dingz.ch) Smart Home Device.",
   "license": "Apache-2.0",
@@ -24,6 +24,7 @@
     "prepublishOnly": "yarn run lint && yarn run build",
     "prepare": "yarn run lint && yarn run build",
     "changelog": "changelog",
+    "release:candidate": "yarn version --prerelease --preid rc --no-git-tag-version && yarn publish --tag next",
     "release:nightly": "yarn version --prerelease --preid nightly --no-git-tag-version && yarn publish --tag nightly",
     "release:major": "yarn run prepare && changelog -M && git add CHANGELOG.md && git commit -m 'ci(docs): updated CHANGELOG.md' && yarn version --major && git push origin && git push origin --tags",
     "release:minor": "yarn run prepare && changelog -m && git add CHANGELOG.md && git commit -m 'ci(docs): updated CHANGELOG.md' && yarn version --minor && git push origin && git push origin --tags",
@@ -41,11 +42,16 @@
     "abort-controller": "^3.0.0",
     "async-mutex": "^0.2.2",
     "axios": "^0.21.0",
+    "axios-debug-log": "^0.8.2",
     "axios-retry": "^3.1.8",
     "body-parser": "^1.19.0",
-    "cockatiel": "^2.0.0",
+    "chalk": "^4.1.0",
+    "cockatiel": "^1.1.1",
     "express": "^4.17.1",
+    "intervals-for-humans": "^1.0.4",
+    "is-valid-host": "^0.0.1",
     "qs": "^6.9.4",
+    "semver": "^7.3.2",
     "simple-color-converter": "^2.1.5"
   },
   "devDependencies": {
@@ -55,8 +61,9 @@
     "@types/express": "^4.17.6",
     "@types/form-data": "^2.5.0",
     "@types/qs": "^6.9.2",
-    "@typescript-eslint/eslint-plugin": "^2.33.0",
-    "@typescript-eslint/parser": "^2.33.0",
+    "@types/semver": "^7.3.4",
+    "@typescript-eslint/eslint-plugin": "^4.8.2",
+    "@typescript-eslint/parser": "^4.8.2",
     "eslint": "^6.8.0",
     "eslint-config-prettier": "^6.11.0",
     "eslint-plugin-prettier": "^3.1.3",
@@ -68,7 +75,7 @@
     "prettier": "2.2.1",
     "rimraf": "^3.0.2",
     "ts-node": "^9.0.0",
-    "typescript": "^3.8.3"
+    "typescript": "^4.1.2"
   },
   "husky": {
     "hooks": {
