@@ -7,6 +7,6 @@ import { DingzDaHomebridgePlatform } from './platform';
 /**
  * This method registers the platform with Homebridge
  */
-export = (api: API) => {
+export = (api: API): void => {
   api.registerPlatform(PLUGIN_NAME, PLATFORM_NAME, DingzDaHomebridgePlatform);
 };
