@@ -4,44 +4,47 @@ import { ButtonId } from './dingzTypes';
 
 // Platform elements
 // EVENT TYPES
-export const enum DingzEvent {
-  UPDATE_INFO = 'updateDingzInfo',
+export const enum PlatformEvent {
+  UPDATE_DEVICE_INFO = 'updateDeviceInfo',
   ACTION = 'deviceAction',
-  STATE_UPDATE = 'stateUpdate',
+  REQUEST_STATE_UPDATE = 'requestStateUpdate',
 }
 
-export declare interface DingzEventBus {
+export declare interface PlatformEventBus {
   on(
-    event: DingzEvent.ACTION,
+    event: PlatformEvent.ACTION,
     listener: (mac: string, action: ButtonAction, battery: number) => void,
   ): this;
   on(
-    event: DingzEvent.ACTION,
+    event: PlatformEvent.ACTION,
     listener: (mac: string, action: ButtonAction, button: ButtonId) => void,
   ): this;
   on(
-    event: DingzEvent.UPDATE_INFO,
+    event: PlatformEvent.UPDATE_DEVICE_INFO,
     listener: (deviceInfo: DeviceInfo) => void,
   ): this;
-  on(event: DingzEvent.STATE_UPDATE, listener: () => void): this;
+  on(event: PlatformEvent.REQUEST_STATE_UPDATE, listener: () => void): this;
 
   emit(
-    event: DingzEvent.ACTION,
+    event: PlatformEvent.ACTION,
     mac: string,
     action: ButtonAction,
     battery: number,
   ): boolean;
   emit(
-    event: DingzEvent.ACTION,
+    event: PlatformEvent.ACTION,
     mac: string,
     action: ButtonAction,
     button: ButtonId,
   ): boolean;
-  emit(event: DingzEvent.UPDATE_INFO, deviceInfo: DeviceInfo): boolean;
-  emit(event: DingzEvent.STATE_UPDATE): boolean;
+  emit(
+    event: PlatformEvent.UPDATE_DEVICE_INFO,
+    deviceInfo: DeviceInfo,
+  ): boolean;
+  emit(event: PlatformEvent.REQUEST_STATE_UPDATE): boolean;
 }
 
-export class DingzEventBus extends EventEmitter {
+export class PlatformEventBus extends EventEmitter {
   constructor() {
     super();
     this.setMaxListeners(20); // Maximum of 20 services
