@@ -8,10 +8,11 @@ import { Policy } from 'cockatiel';
 import { Mutex } from 'async-mutex';
 
 import { DingzDaHomebridgePlatform } from './platform';
-import { MyStromDeviceInfo, MyStromPIRReport } from './util/myStromTypes';
-import { ButtonAction, DeviceInfo } from './util/commonTypes';
-import { DeviceNotReachableError } from './util/errors';
-import { DingzEvent } from './util/dingzEventBus';
+import { MyStromDeviceInfo, MyStromPIRReport } from './lib/myStromTypes';
+import { ButtonAction } from './lib/commonTypes';
+import { DeviceNotReachableError } from './lib/errors';
+import { PlatformEvent } from './lib/platformEventBus';
+import { DingzDaBaseAccessory } from './lib/dingzDaBaseAccessory';
 
 // Policy for long running tasks, retry every hour
 const retrySlow = Policy.handleAll()
@@ -24,7 +25,7 @@ const retrySlow = Policy.handleAll()
  * An instance of this class is created for each accessory your platform registers
  * Each accessory may expose multiple services of different service types.
  */
-export class MyStromPIRAccessory {
+export class MyStromPIRAccessory extends DingzDaBaseAccessory {
   private readonly mutex = new Mutex();
   private services: Service[] = [];
   private motionService: Service;
@@ -32,9 +33,7 @@ export class MyStromPIRAccessory {
   private lightService: Service;
 
   // Eventually replaced by:
-  private device: DeviceInfo;
   private mystromDeviceInfo: MyStromDeviceInfo;
-  private baseUrl: string;
 
   private pirState = {
     motion: false,
@@ -43,15 +42,14 @@ export class MyStromPIRAccessory {
   } as MyStromPIRReport;
 
   constructor(
-    private readonly platform: DingzDaHomebridgePlatform,
-    private readonly accessory: PlatformAccessory,
+    private readonly _platform: DingzDaHomebridgePlatform,
+    private readonly _accessory: PlatformAccessory,
   ) {
+    super(_platform, _accessory);
     // Set Base URL
-    this.device = this.accessory.context.device;
     this.mystromDeviceInfo = this.device.hwInfo as MyStromDeviceInfo;
-    this.baseUrl = `http://${this.device.address}`;
 
-    this.platform.log.debug(
+    this.log.debug(
       'Setting informationService Characteristics ->',
       this.device.model,
     );
@@ -81,7 +79,7 @@ export class MyStromPIRAccessory {
 
     // get the LightBulb service if it exists, otherwise create a new LightBulb service
     // you can create multiple services for each accessory
-    this.platform.log.info('Create Motion Sensor -> ', this.device.name);
+    this.log.info('Create Motion Sensor');
 
     this.temperatureService =
       this.accessory.getService(this.platform.Service.TemperatureSensor) ??
@@ -120,67 +118,19 @@ export class MyStromPIRAccessory {
       'Motion',
     );
 
-    // Only check for motion if we have a PIR and set the Interval
-    setInterval(() => {
-      this.getDeviceReport()
-        .then((report) => {
-          if (report) {
-            // If we are in motion polling mode, update motion from poller
-            if (this.platform.config.motionPoller ?? true) {
-              this.platform.log.info(
-                'Motion POLLING of',
-                this.device.name,
-                'enabled',
-              );
-              const isMotion: boolean = report.motion;
-              // Only update if motionService exists *and* if there's a change in motion'
-              if (this.pirState.motion !== isMotion) {
-                this.platform.log.debug('Motion Update from POLLER');
-                this.pirState.motion = isMotion;
-                this.motionService.updateCharacteristic(
-                  this.platform.Characteristic.MotionDetected,
-                  this.pirState.motion,
-                );
-              }
-            }
-
-            // Update temperature and light in any case
-            this.pirState.temperature = report.temperature;
-            this.temperatureService.updateCharacteristic(
-              this.platform.Characteristic.CurrentTemperature,
-              this.pirState.temperature,
-            );
-
-            this.pirState.light = report.light ?? 0;
-            this.lightService.updateCharacteristic(
-              this.platform.Characteristic.CurrentAmbientLightLevel,
-              this.pirState.light,
-            );
-          } else {
-            throw new DeviceNotReachableError(
-              `Device can not be reached ->
-              ${this.device.name}-> ${this.device.address}`,
-            );
-          }
-        })
-        .catch((e: Error) => {
-          this.platform.log.error(
-            'Error -> unable to fetch DeviceMotion data',
-            e.name,
-            e.toString(),
-          );
-        });
-    }, 2000); // Shorter term updates for motion sensor
+    // Subscribe to the REQUEST_STATE_UPDATE event
+    this.platform.eb.on(
+      PlatformEvent.REQUEST_STATE_UPDATE,
+      this.getDeviceStateUpdate.bind(this),
+    );
 
     if (!(this.platform.config.motionPoller ?? true)) {
       // Implement *push* event handling
-      this.platform.eb.on(DingzEvent.ACTION, (mac, action) => {
-        this.platform.log.debug(`Processing DingzEvent.ACTION ${action}`);
+      this.platform.eb.on(PlatformEvent.ACTION, (mac, action) => {
+        this.log.debug(`Processing DingzEvent.ACTION ${action}`);
 
         if (mac === this.device.mac) {
-          this.platform.log.debug(
-            `Motion detected by ${this.device.name} (${this.device.mac}) pressed -> ${action}`,
-          );
+          this.log.debug(`Motion detected -> ${action}`);
           let isMotion: boolean | undefined;
           switch (action) {
             case ButtonAction.PIR_MOTION_STOP:
@@ -192,9 +142,9 @@ export class MyStromPIRAccessory {
               break;
           }
 
-          this.platform.log.debug('Motion Update from PUSH');
+          this.log.debug('Motion Update from PUSH');
           this.pirState.motion = isMotion;
-          this.motionService.updateCharacteristic(
+          this.motionService.setCharacteristic(
             this.platform.Characteristic.MotionDetected,
             this.pirState.motion,
           );
@@ -212,19 +162,58 @@ export class MyStromPIRAccessory {
     });
   }
 
+  // Get updated device info and update the corresponding values
+  protected getDeviceStateUpdate(): Promise<void> {
+    return this.getDeviceReport()
+      .then((report) => {
+        if (report) {
+          // If we are in motion polling mode, update motion from poller
+          // TODO: remove this -- doesn't make sense at all
+          if (this.platform.config.motionPoller ?? true) {
+            this.log.info('Motion POLLING of', this.device.name, 'enabled');
+            const isMotion: boolean = report.motion;
+            // Only update if motionService exists *and* if there's a change in motion'
+            if (this.pirState.motion !== isMotion) {
+              this.log.debug('Motion Update from POLLER');
+              this.pirState.motion = isMotion;
+              this.motionService
+                .getCharacteristic(this.platform.Characteristic.MotionDetected)
+                .updateValue(this.pirState.motion);
+            }
+          }
+
+          // Update temperature and light in any case
+          this.pirState.temperature = report.temperature;
+          this.temperatureService
+            .getCharacteristic(this.platform.Characteristic.CurrentTemperature)
+            .updateValue(this.pirState.temperature);
+
+          this.pirState.light = report.light ?? 0;
+          this.lightService
+            .getCharacteristic(
+              this.platform.Characteristic.CurrentAmbientLightLevel,
+            )
+            .updateValue(this.pirState.light);
+
+          return Promise.resolve();
+        } else {
+          return Promise.reject(new DeviceNotReachableError());
+        }
+      })
+      .catch((e: Error) => {
+        return Promise.reject(e);
+      });
+  }
+
   /**
    * Handle Handle the "GET" requests from HomeKit
    * to get the current value of the "Ambient Light Level" characteristic
    */
   private getLightLevel(callback: CharacteristicGetCallback) {
     const light: number = this.pirState?.light ?? 42;
-    this.platform.log.debug(
-      'Get Characteristic Ambient Light Level ->',
-      light,
-      ' lux',
-    );
+    this.log.debug('Get Characteristic Ambient Light Level ->', light, ' lux');
 
-    callback(null, light);
+    callback(this.reachabilityState, light);
   }
 
   /**
@@ -233,13 +222,9 @@ export class MyStromPIRAccessory {
    */
   private getTemperature(callback: CharacteristicGetCallback) {
     const temperature: number = this.pirState?.temperature;
-    this.platform.log.debug(
-      'Get Characteristic Temperature ->',
-      temperature,
-      '° C',
-    );
+    this.log.debug('Get Characteristic Temperature ->', temperature, '° C');
 
-    callback(null, temperature);
+    callback(this.reachabilityState, temperature);
   }
 
   /**
@@ -249,21 +234,14 @@ export class MyStromPIRAccessory {
   private getMotionDetected(callback: CharacteristicGetCallback) {
     // set this to a valid value for MotionDetected
     const isMotion = this.pirState.motion;
-    callback(null, isMotion);
+    callback(this.reachabilityState, isMotion);
   }
 
   private async getDeviceReport(): Promise<MyStromPIRReport> {
     const getSensorsUrl = `${this.baseUrl}/api/v1/sensors`;
-    const release = await this.mutex.acquire();
-    try {
-      return await this.platform.fetch({
-        url: getSensorsUrl,
-        returnBody: true,
-        token: this.device.token,
-      });
-    } finally {
-      release();
-    }
+    return await this.request.get(getSensorsUrl).then((response) => {
+      return response.data;
+    });
   }
 
   /*
@@ -271,9 +249,6 @@ export class MyStromPIRAccessory {
    * Typical this only ever happens at the pairing process.
    */
   identify(): void {
-    this.platform.log.debug(
-      'Identify! -> Who am I? I am',
-      this.accessory.displayName,
-    );
+    this.log.debug('Identify! -> Who am I? I am', this.accessory.displayName);
   }
 }
